Fix Mallets movement and make Background.
But there are some problems with running on my device. please run it and if it has bugs, please tell me to fix it.
Maybe I will add some script to the background canvas. but I guess it works well rn.
the mallet will be turned by touching the screen and the background moves with camera.